+++ src/qtils/bytes.hpp
@@ -11,8 +11,6 @@
 #include <span>
 #include <vector>
 
-#include <qtils/assert.hpp>
-
 namespace qtils {
   using Bytes = std::vector<uint8_t>;
   template <size_t N>
@@ -36,4 +34,7 @@ namespace qtils {
   template <typename T>
   concept AsBytes = requires(const T &t) { BytesIn{t}; };
 
+  inline Bytes asVec(BytesIn v) {
+    return {v.begin(), v.end()};
+  }
 }  // namespace qtils
